- Adding pyside6 compatibility

+++ functions.py
@@ -354,7 +354,7 @@ def f_axis_idx_map(freqs, freq_range=None):
         sign = np.sign(freq_range[1] - freq_range[0])
         f0_idx = int(np.argmin(np.abs(freqs - freq_range[0])))
         f1_idx = int(np.argmin(np.abs(freqs - freq_range[1])))
-        f_idx = np.arange(f0_idx, f1_idx + 1, sign)
+        f_idx = np.arange(f0_idx, f1_idx + 1, sign, dtype=int)
     elif freq_range == "full":
         f_idx = np.ones_like(freqs, dtype=bool)
     else:
